Adds blind voting publish results setting Fixes report page not working
