+++ Source/Extension.Voting/src/adminPage/adminPage.html
@@ -128,12 +128,19 @@
                         </div>
                     </div>
 
-                    <div class="row">                       
-                        <div class="form-group col-sm-12">
+                    <div class="row">
+                        <div class="form-group col-sm-6">
                             <label>
                                 <input type="checkbox" name="isBlindVotingEnabled" v-model="actualVoting.isBlindVotingEnabled" />
                                 <span data-toggle="tooltip"
                                     title="If blind voting is enabled, users can only see their own votes but not votes of other users.">Enable blind voting</span>
+                            </label>
+                        </div>
+                        <div class="form-group col-sm-6">
+                            <label>
+                                <input type="checkbox" name="isBlindVotingPublishEnabled" v-model="actualVoting.isBlindVotingPublishEnabled" />
+                                <span data-toggle="tooltip"
+                                    title="If blind voting publish results is enabled, users can see the final result when the voting is stopped, paused or has ended.">Enable blind voting publish results</span>
                             </label>
                         </div>
                     </div>
