+++ Source/Extension.Voting/src/votingPage/votingPage.html
@@ -91,10 +91,10 @@
                     </p>
                 </div>
 
-                <div class="col-sm-12" v-if="actualVoting.isBlindVotingEnabled || actualVoting.cannotVoteForAssignedWorkItems || actualVoting.cannotVoteForOwnWorkItems">
+                <div class="col-sm-12" v-if="isBlindVotingEnabled || actualVoting.cannotVoteForAssignedWorkItems || actualVoting.cannotVoteForOwnWorkItems">
                     <p>
                         <span class="bowtie-icon bowtie-status-info"></span>
-                        <span v-if="actualVoting.isBlindVotingEnabled">Blind voting is enabled. You can only see your own votes. </span>
+                        <span v-if="isBlindVotingEnabled">Blind voting is enabled. You can only see your own votes. </span>
                         <span v-if="actualVoting.cannotVoteForAssignedWorkItems">You cannot vote for work items which are assigned to you. </span>
                         <span v-if="actualVoting.cannotVoteForOwnWorkItems">You cannot vote for work items you created. </span>
                     </p>
